ChatGenieMarketplace/requirements.txt
# Streamlit dependencies
streamlit==1.46.1
streamlit-float==0.3.5
altair>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Databricks dependencies
databricks-sdk>=0.56.0
databricks-sql-connector>=4.0.5

# Database connections (for AppFrontEnd data module)
psycopg2-binary>=2.9.10

# Data processing
sqlparse==0.5.3

# Optional: Keep Dash dependencies commented in case needed for backup
# dash==2.18.2
# dash-bootstrap-components==1.6.0
# dash-core-components==2.0.0
# dash-html-components==2.0.0
# dash-iconify==0.1.2
# dash-leaflet==1.0.15
# dash-table==5.0.0
# dash_ag_grid==31.3.0
# dash_mantine_components==0.15.3
# backoff==2.2.1